Move permissions to crossbow-android
