+++ Cargo.toml
@@ -15,7 +15,6 @@ crossbundle-derive = { path = "crossbundle/derive", version = "0.1.3" }
 crossbundle-tools = { path = "crossbundle/tools", version = "0.1.3", optional = true }
 
 crossbow-ads = { path = "plugins/ads", version = "0.1.3", optional = true }
-crossbow-permissions = { path = "plugins/permissions", version = "0.1.3", optional = true }
 
 [target.'cfg(target_os = "android")'.dependencies]
 ndk-glue = "0.6.2"
