+++ examples/macroquad-permissions/Cargo.toml
@@ -5,7 +5,7 @@ authors = ["DodoRare Team <support@dodorare.com>"]
 edition = "2021"
 
 [dependencies]
-crossbow = { version = "0.1.3", path = "../../", features = ["crossbow-permissions"] }
+crossbow = { version = "0.1.3", path = "../../" }
 crossbow-android = { path = "../../platform/android" }
 log = "0.4"
 anyhow = "1.0"
