+++ platform/android/Cargo.toml
@@ -10,7 +10,13 @@ keywords = ["crossbow", "android", "port"]
 readme = "README.md"
 
 [dependencies]
+thiserror = "1.0"
+displaydoc = "0.2"
+anyhow = "1.0"
 jni = "0.19"
+ndk-context = "0.1"
+# ndk-glue = "0.6.2"
+# ndk = "0.6"
 
 [package.metadata]
 app_name = "crossbow_android"
