+++ platform/android/src/lib.rs
@@ -1,3 +1,7 @@
+#![cfg(target_os = "android")]
+
+pub mod permission;
+
 use jni::{
     objects::{JClass, JString},
     sys::jboolean,
