+++ platform/android/src/permission/error.rs
@@ -7,10 +7,7 @@ pub type Result<T> = std::result::Result<T, PermissionError>;
 /// `Crossbow-permissions` error type.
 #[derive(Display, Debug, Error)]
 pub enum PermissionError {
-    /// Requesting permission on the wrong platform
-    PermissionWrongPlatform,
     /// Rust Jni library error
-    #[cfg(target_os = "android")]
     Jni(jni::errors::Error),
     /// Anyhow library errors
     Anyhow(anyhow::Error),
