+++ platform/android/src/permission/mod.rs
@@ -1,0 +1,8 @@
+pub mod error;
+pub mod request_permission;
+pub mod types;
+
+pub mod prelude {
+    pub use super::request_permission::*;
+    pub use super::types::android::*;
+}
