+++ platform/android/src/permission/types/android_permission_group.rs
