+++ platform/android/src/permission/types/consts.rs
