+++ platform/android/src/permission/types/mod.rs
