--- plugins/permissions/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "crossbow-permissions"
-version = "0.1.3"
-edition = "2021"
-authors = ["DodoRare Team <support@dodorare.com>"]
-description = "Cross-Platform Rust Toolkit for Games 🏹"
-repository = "https://github.com/dodorare/crossbow"
-license = "Apache-2.0"
-keywords = ["permissions", "android", "ios"]
-readme = "README.md"
-
-[dependencies]
-thiserror = "1.0"
-displaydoc = "0.1"
-anyhow = "1.0"
-
-[target.'cfg(target_os = "android")'.dependencies]
-ndk-context = "0.1"
-ndk-glue = "0.6.2"
-jni = "0.19"
-ndk = "0.6"
-
-[features]
-default = ["full-permissions"]
-full-permissions = []
