--- plugins/permissions/README.md
@@ -1,3 +1,0 @@
-# Crossbow Permissions
-
-The main purpose of this plugin is to give cross-platform access to permissions.
