+++ src/lib.rs
@@ -6,8 +6,5 @@ pub use crossbow_ads;
 #[cfg(feature = "crossbundle-tools")]
 pub use crossbundle_tools;
 
-#[cfg(feature = "crossbow-permissions")]
-pub use crossbow_permissions;
-
 #[cfg(target_os = "android")]
 pub use ndk_glue;
